Added Catenaccio Football Manager
The game runs in iOS and Android native apps into the system webview, using Howler we solved the compatibility issues. If you want to add it in showcase with images we can send you the 2x and 1x image at the resolution you need it. I noticed that a few links are not reachable, if you want we can take one of that spots. Contact us: info [at] catenacc.io

Great job!